buffs v0.1
working buff 'system'

+++ Buffs/SkillBuff.java
@@ -4,7 +4,7 @@ import Player.Character;
 
 public class SkillBuff extends Buff{
 
-	public SkillBuff(Character owner, short buffId, short buffLength, short buffValue){
+	public SkillBuff(Character owner, short buffId, long buffLength, short buffValue){
 		super(owner, buffId, buffLength, buffValue);
 	}
